Update quizQuestions.js
this is the template for the questions we will use, inside the component directory is the functions to add functions and in the app.js folder you'll see the shuffle function. i added a swap function to create a random swapping of questions and answers, etc. for now, we need to figure out the questions we want to add. also, we will need style our application more.